requirements.txt
anyio==3.7.1
asgiref==3.8.1
bcrypt==3.2.0
cffi==1.16.0
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
motor==2.5.1
optional==0.0.1
packaging==24.0
passlib==1.7.4
pycparser==2.22
pydantic==1.10.15
PyJWT==2.3.0
pymongo==3.13.0
python-decouple==3.3
six==1.16.0
sniffio==1.3.1
starlette==0.16.0
typing_extensions==4.11.0
uvicorn==0.14.0
